+++ mayan/apps/events/links.py
@@ -5,8 +5,14 @@ from mayan.apps.navigation.classes import Link
 from mayan.apps.navigation.utils import get_content_type_kwargs_factory
 
 from .icons import (
-    icon_events_list, icon_events_for_object,
-    icon_event_types_subscriptions_list, icon_notification_mark_read,
+    icon_current_user_events,
+    icon_current_user_events_export,
+    icon_event_types_subscriptions_list,
+    icon_events_for_object,
+    icon_events_for_object_export,
+    icon_events_list,
+    icon_events_list_export,
+    icon_notification_mark_read,
     icon_notification_mark_read_all,
     icon_object_event_types_user_subcriptions_list,
     icon_user_notifications_list
@@ -25,8 +31,12 @@ def get_unread_notification_count(context):
 
 
 link_current_user_events = Link(
-    icon=icon_events_list, text=_('My events'),
+    icon=icon_current_user_events, text=_('My events'),
     view='events:current_user_events'
+)
+link_current_user_events_export = Link(
+    icon=icon_current_user_events_export, text=_('Export'),
+    view='events:current_user_events_export'
 )
 link_events_details = Link(
     text=_('Events'), view='events:events_list'
@@ -35,10 +45,20 @@ link_events_for_object = Link(
     icon=icon_events_for_object,
     kwargs=get_content_type_kwargs_factory(variable_name='resolved_object'),
     permissions=(permission_events_view,), text=_('Events'),
-    view='events:events_for_object',
+    view='events:events_for_object'
+)
+link_events_list_export = Link(
+    icon=icon_events_for_object_export,
+    kwargs=get_content_type_kwargs_factory(variable_name='resolved_object'),
+    permissions=(permission_events_view,), text=_('Export'),
+    view='events:events_for_object_export'
 )
 link_events_list = Link(
     icon=icon_events_list, text=_('Events'), view='events:events_list'
+)
+link_events_list_export = Link(
+    icon=icon_events_list_export, text=_('Export'),
+    view='events:events_list_export'
 )
 link_event_types_subscriptions_list = Link(
     icon=icon_event_types_subscriptions_list,