+++ mayan/apps/events/tests/mixins.py
@@ -103,26 +103,25 @@ class EventViewTestMixin:
 
 class EventsExportViewTestMixin:
     def _request_test_current_user_events_export_view(self):
-        return self.get(
+        return self.post(
             viewname='events:current_user_events_export'
         )
 
     def _request_test_events_by_verb_export_view(self):
-        return self.get(
+        return self.post(
             viewname='events:events_by_verb_export', kwargs={
                 'verb': self.test_event_type.id
             }
         )
 
     def _request_test_events_list_export_view(self):
-        return self.get(viewname='events:events_list_export')
+        return self.post(viewname='events:events_list_export')
 
     def _request_events_for_object_export_view(self):
-        return self.get(
+        return self.post(
             viewname='events:events_for_object_export',
             kwargs=self.view_arguments
         )
-
 
 
 class NotificationTestMixin(