+++ mayan/apps/events/tests/test_views.py
@@ -1,8 +1,14 @@
 from django.contrib.contenttypes.models import ContentType
 
+from mayan.apps.acls.classes import ModelPermission
 from mayan.apps.documents.tests.base import GenericDocumentViewTestCase
+from mayan.apps.messaging.events import event_message_created
+from mayan.apps.messaging.models import Message
 from mayan.apps.testing.tests.base import GenericViewTestCase
+from mayan.apps.storage.events import event_download_file_created
+from mayan.apps.storage.models import DownloadFile
 
+from ..events import event_events_exported
 from ..models import Notification
 from ..permissions import permission_events_export, permission_events_view
 
@@ -151,6 +157,14 @@ class EventExportViewTestCase(
             'object_id': self.test_object.pk
         }
 
+        ModelPermission.register(
+            model=self.test_object._meta.model, permissions=(
+                permission_events_export,
+            )
+        )
+
+        self._clear_events()
+
     def create_test_event(self, **kwargs):
         self.test_action = self.test_event_type.commit(**kwargs)
         self.test_actions.append(self.test_action)
@@ -159,17 +173,71 @@ class EventExportViewTestCase(
         self.create_test_event(target=self.test_object)
 
         response = self._request_test_events_list_export_view()
-        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.status_code, 302)
+
+        test_download_file = DownloadFile.objects.first()
+        test_message = Message.objects.first()
+
+        events = self._get_test_events()
+        # Test object creation + 3 for the test
+        self.assertEqual(events.count(), 4)
+
+        self.assertEqual(events[1].action_object, None)
+        self.assertEqual(events[1].actor, self._test_case_user)
+        self.assertEqual(events[1].target, test_download_file)
+        self.assertEqual(events[1].verb, event_download_file_created.id)
+
+        self.assertEqual(events[2].action_object, None)
+        self.assertEqual(events[2].actor, self._test_case_user)
+        self.assertEqual(events[2].target, test_download_file)
+        self.assertEqual(events[2].verb, event_events_exported.id)
+
+        self.assertEqual(events[3].action_object, None)
+        self.assertEqual(events[3].actor, test_message)
+        self.assertEqual(events[3].target, test_message)
+        self.assertEqual(events[3].verb, event_message_created.id)
+
+        with test_download_file.open() as file_object:
+            self.assertTrue(
+                str(self.test_object).encode() not in file_object.read()
+            )
 
     def test_events_list_export_view_with_access(self):
         self.create_test_event(target=self.test_object)
 
         self.grant_access(
-            obj=self.test_object, permission=permission_events_view
+            obj=self.test_object, permission=permission_events_export
         )
 
         response = self._request_test_events_list_export_view()
-        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.status_code, 302)
+
+        test_download_file = DownloadFile.objects.first()
+        test_message = Message.objects.first()
+
+        events = self._get_test_events()
+        # Test object creation + access grant + 3 for the test
+        self.assertEqual(events.count(), 5)
+
+        self.assertEqual(events[2].action_object, None)
+        self.assertEqual(events[2].actor, self._test_case_user)
+        self.assertEqual(events[2].target, test_download_file)
+        self.assertEqual(events[2].verb, event_download_file_created.id)
+
+        self.assertEqual(events[3].action_object, None)
+        self.assertEqual(events[3].actor, self._test_case_user)
+        self.assertEqual(events[3].target, test_download_file)
+        self.assertEqual(events[3].verb, event_events_exported.id)
+
+        self.assertEqual(events[4].action_object, None)
+        self.assertEqual(events[4].actor, test_message)
+        self.assertEqual(events[4].target, test_message)
+        self.assertEqual(events[4].verb, event_message_created.id)
+
+        with test_download_file.open() as file_object:
+            self.assertTrue(
+                str(self.test_object).encode() in file_object.read()
+            )
 
     def test_events_for_object_export_view_no_permission(self):
         self.create_test_event(
@@ -177,7 +245,34 @@ class EventExportViewTestCase(
         )
 
         response = self._request_events_for_object_export_view()
-        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.status_code, 302)
+
+        test_download_file = DownloadFile.objects.first()
+        test_message = Message.objects.first()
+
+        events = self._get_test_events()
+        # Test object creation + 3 for the test
+        self.assertEqual(events.count(), 4)
+
+        self.assertEqual(events[1].action_object, None)
+        self.assertEqual(events[1].actor, self._test_case_user)
+        self.assertEqual(events[1].target, test_download_file)
+        self.assertEqual(events[1].verb, event_download_file_created.id)
+
+        self.assertEqual(events[2].action_object, None)
+        self.assertEqual(events[2].actor, self._test_case_user)
+        self.assertEqual(events[2].target, test_download_file)
+        self.assertEqual(events[2].verb, event_events_exported.id)
+
+        self.assertEqual(events[3].action_object, None)
+        self.assertEqual(events[3].actor, test_message)
+        self.assertEqual(events[3].target, test_message)
+        self.assertEqual(events[3].verb, event_message_created.id)
+
+        with test_download_file.open() as file_object:
+            self.assertTrue(
+                str(self.test_object).encode() not in file_object.read()
+            )
 
     def test_events_for_object_export_view_with_access(self):
         self.create_test_event(
@@ -185,11 +280,38 @@ class EventExportViewTestCase(
         )
 
         self.grant_access(
-            obj=self.test_object, permission=permission_events_view
+            obj=self.test_object, permission=permission_events_export
         )
 
         response = self._request_events_for_object_export_view()
-        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.status_code, 302)
+
+        test_download_file = DownloadFile.objects.first()
+        test_message = Message.objects.first()
+
+        events = self._get_test_events()
+        # Test object creation + access grant + 3 for the test
+        self.assertEqual(events.count(), 5)
+
+        self.assertEqual(events[2].action_object, None)
+        self.assertEqual(events[2].actor, self._test_case_user)
+        self.assertEqual(events[2].target, test_download_file)
+        self.assertEqual(events[2].verb, event_download_file_created.id)
+
+        self.assertEqual(events[3].action_object, None)
+        self.assertEqual(events[3].actor, self._test_case_user)
+        self.assertEqual(events[3].target, test_download_file)
+        self.assertEqual(events[3].verb, event_events_exported.id)
+
+        self.assertEqual(events[4].action_object, None)
+        self.assertEqual(events[4].actor, test_message)
+        self.assertEqual(events[4].target, test_message)
+        self.assertEqual(events[4].verb, event_message_created.id)
+
+        with test_download_file.open() as file_object:
+            self.assertTrue(
+                str(self.test_object).encode() in file_object.read()
+            )
 
     def test_events_by_verb_export_view_no_permission(self):
         self.create_test_event(
@@ -197,7 +319,34 @@ class EventExportViewTestCase(
         )
 
         response = self._request_test_events_by_verb_export_view()
-        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.status_code, 302)
+
+        test_download_file = DownloadFile.objects.first()
+        test_message = Message.objects.first()
+
+        events = self._get_test_events()
+        # Test object creation + 3 for the test
+        self.assertEqual(events.count(), 4)
+
+        self.assertEqual(events[1].action_object, None)
+        self.assertEqual(events[1].actor, self._test_case_user)
+        self.assertEqual(events[1].target, test_download_file)
+        self.assertEqual(events[1].verb, event_download_file_created.id)
+
+        self.assertEqual(events[2].action_object, None)
+        self.assertEqual(events[2].actor, self._test_case_user)
+        self.assertEqual(events[2].target, test_download_file)
+        self.assertEqual(events[2].verb, event_events_exported.id)
+
+        self.assertEqual(events[3].action_object, None)
+        self.assertEqual(events[3].actor, test_message)
+        self.assertEqual(events[3].target, test_message)
+        self.assertEqual(events[3].verb, event_message_created.id)
+
+        with test_download_file.open() as file_object:
+            self.assertTrue(
+                str(self.test_object).encode() not in file_object.read()
+            )
 
     def test_events_by_verb_view_export_with_access(self):
         self.create_test_event(
@@ -205,11 +354,38 @@ class EventExportViewTestCase(
         )
 
         self.grant_access(
-            obj=self.test_object, permission=permission_events_view
+            obj=self.test_object, permission=permission_events_export
         )
 
         response = self._request_test_events_by_verb_export_view()
-        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.status_code, 302)
+
+        test_download_file = DownloadFile.objects.first()
+        test_message = Message.objects.first()
+
+        events = self._get_test_events()
+        # Test object creation + access grant + 3 for the test
+        self.assertEqual(events.count(), 5)
+
+        self.assertEqual(events[2].action_object, None)
+        self.assertEqual(events[2].actor, self._test_case_user)
+        self.assertEqual(events[2].target, test_download_file)
+        self.assertEqual(events[2].verb, event_download_file_created.id)
+
+        self.assertEqual(events[3].action_object, None)
+        self.assertEqual(events[3].actor, self._test_case_user)
+        self.assertEqual(events[3].target, test_download_file)
+        self.assertEqual(events[3].verb, event_events_exported.id)
+
+        self.assertEqual(events[4].action_object, None)
+        self.assertEqual(events[4].actor, test_message)
+        self.assertEqual(events[4].target, test_message)
+        self.assertEqual(events[4].verb, event_message_created.id)
+
+        with test_download_file.open() as file_object:
+            self.assertTrue(
+                str(self.test_object).encode() in file_object.read()
+            )
 
     def test_current_user_events_export_view_no_permission(self):
         self.create_test_event(
@@ -218,8 +394,35 @@ class EventExportViewTestCase(
 
         response = self._request_test_current_user_events_export_view()
         self.assertNotContains(
-            response=response, text=str(self.test_event_type), status_code=200
+            response=response, text=str(self.test_event_type), status_code=302
         )
+
+        test_download_file = DownloadFile.objects.first()
+        test_message = Message.objects.first()
+
+        events = self._get_test_events()
+        # Test object creation + 3 for the test
+        self.assertEqual(events.count(), 4)
+
+        self.assertEqual(events[1].action_object, None)
+        self.assertEqual(events[1].actor, self._test_case_user)
+        self.assertEqual(events[1].target, test_download_file)
+        self.assertEqual(events[1].verb, event_download_file_created.id)
+
+        self.assertEqual(events[2].action_object, None)
+        self.assertEqual(events[2].actor, self._test_case_user)
+        self.assertEqual(events[2].target, test_download_file)
+        self.assertEqual(events[2].verb, event_events_exported.id)
+
+        self.assertEqual(events[3].action_object, None)
+        self.assertEqual(events[3].actor, test_message)
+        self.assertEqual(events[3].target, test_message)
+        self.assertEqual(events[3].verb, event_message_created.id)
+
+        with test_download_file.open() as file_object:
+            self.assertTrue(
+                str(self.test_object).encode() not in file_object.read()
+            )
 
     def test_current_user_events_export_view_with_access(self):
         self.create_test_event(
@@ -227,11 +430,38 @@ class EventExportViewTestCase(
         )
 
         self.grant_access(
-            obj=self.test_object, permission=permission_events_view
+            obj=self.test_object, permission=permission_events_export
         )
 
         response = self._request_test_current_user_events_export_view()
-        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.status_code, 302)
+
+        test_download_file = DownloadFile.objects.first()
+        test_message = Message.objects.first()
+
+        events = self._get_test_events()
+        # Test object creation + access grant + 3 for the test
+        self.assertEqual(events.count(), 5)
+
+        self.assertEqual(events[2].action_object, None)
+        self.assertEqual(events[2].actor, self._test_case_user)
+        self.assertEqual(events[2].target, test_download_file)
+        self.assertEqual(events[2].verb, event_download_file_created.id)
+
+        self.assertEqual(events[3].action_object, None)
+        self.assertEqual(events[3].actor, self._test_case_user)
+        self.assertEqual(events[3].target, test_download_file)
+        self.assertEqual(events[3].verb, event_events_exported.id)
+
+        self.assertEqual(events[4].action_object, None)
+        self.assertEqual(events[4].actor, test_message)
+        self.assertEqual(events[4].target, test_message)
+        self.assertEqual(events[4].verb, event_message_created.id)
+
+        with test_download_file.open() as file_object:
+            self.assertTrue(
+                str(self.test_object).encode() in file_object.read()
+            )
 
 
 class NotificationViewTestCase(