+++ mayan/apps/events/views/event_views.py
@@ -1,6 +1,6 @@
 from django.utils.translation import ugettext_lazy as _
 
-from actstream.models import Action, any_stream
+from actstream.models import Action, actor_stream, any_stream
 
 from mayan.apps.views.generics import SingleObjectListView
 from mayan.apps.views.mixins import ExternalContentTypeObjectViewMixin
@@ -55,6 +55,9 @@ class ObjectEventListView(ExternalContentTypeObjectViewMixin, EventListBaseView)
 class CurrentUserEventListView(ObjectEventListView):
     def get_external_object(self):
         return self.request.user
+
+    def get_source_queryset(self):
+        return actor_stream(obj=self.external_object)
 
 
 class VerbEventListView(EventListBaseView):